Completely reworked UI and State Transition
Every state ever has been reworked to match on how my vision 2.0 will be.

The drawbacks of doing this memory usage, we have a SEVERE memory usage problem but idc lol. I would probably add an option to disable these later on, but for now just this.

Icons is still using placeholder images, ill replace them when i have the assets.

Next up on the list is still freeplay, but score display then we would probably move on to more UI stuff