Very early Characer Selection menu
For now, it just overrides the SONG.player1, in the future i think i will improvise this, because if theres a character that is not boyfriend by default, it would just override to the selected character

in the future as well, these character that you select will have a cover of its own for the song (replacing boyfriend), some songs might be impossible to do though because they might take alot of effort.

Thats why i think i would want to make a seperate minimal settings that only could effect a current playthrough, after that playthrough is finished, its gonna be reset to defaults. The setting will include the custom cover character thingy, idk will improvise later.